+++ src/core/client/systems/messenger.ts
@@ -2,11 +2,16 @@ import * as alt from 'alt-client';
 
 import { MESSENGER_EVENTS } from '@AthenaShared/enums/messenger';
 import { SYSTEM_EVENTS } from '@AthenaShared/enums/system';
+import { MessageCommand } from '@AthenaShared/interfaces/messageCommand';
 
+export type MessageInfo = { timestamp: number; msg: string };
 type MessageCallback = (msg: string) => void;
+type HistoryCallback = (msgs: Array<MessageInfo>) => void;
 
 const callbacks: Array<MessageCallback> = [];
-const messages: Array<{ timestamp: number; msg: string }> = [];
+const historyCallbacks: Array<HistoryCallback> = [];
+const messages: Array<MessageInfo> = [];
+let commands: Array<Omit<MessageCommand<alt.Player>, 'callback'>> = [];
 
 export const MessengerSystem = {
     /**
@@ -16,8 +21,18 @@ export const MessengerSystem = {
      *
      * @param {MessageCallback} callback
      */
-    register(callback: MessageCallback) {
+    registerMessageCallback(callback: MessageCallback) {
         callbacks.push(callback);
+    },
+    /**
+     * Register a callback that handles messages.
+     * The messages from other clients, and Athena itself will be pushed through all callbacks registered.
+     * Useful for plugin creators.
+     *
+     * @param {HistoryCallback} callback
+     */
+    registerHistoryCallback(callback: HistoryCallback) {
+        historyCallbacks.push(callback);
     },
     /**
      * Emits a message to all callbacks.
@@ -31,13 +46,18 @@ export const MessengerSystem = {
             messages.pop();
         }
 
-        if (callbacks.length <= 0) {
+        if (callbacks.length >= 1) {
+            for (let cb of callbacks) {
+                cb(msg);
+            }
+        } else {
             alt.log(msg);
-            return;
         }
 
-        for (let cb of callbacks) {
-            cb(msg);
+        if (historyCallbacks.length >= 1) {
+            for (let cb of historyCallbacks) {
+                cb(messages);
+            }
         }
     },
     /**
@@ -58,8 +78,29 @@ export const MessengerSystem = {
     send(msg: string) {
         alt.emitServer(MESSENGER_EVENTS.TO_SERVER.MESSAGE, msg);
     },
+
+    /**
+     * Populates the local command list for the client.
+     * Which can be used to
+     *
+     * @param {Array<Omit<MessageCommand<alt.Player>, 'callback'>>} commands
+     */
+    setCommands(_commands: Array<Omit<MessageCommand<alt.Player>, 'callback'>>) {
+        alt.log(`~g~Populated command list with ${_commands.length} commands.`);
+        commands = _commands;
+    },
+
+    /**
+     * Get the commands that this client has permission for.
+     *
+     * @return {Array<Omit<MessageCommand<alt.Player>, 'callback'>>}
+     */
+    getCommands(): Array<Omit<MessageCommand<alt.Player>, 'callback'>> {
+        return commands;
+    },
 };
 
 alt.onServer(SYSTEM_EVENTS.TICKS_START, () => {
     alt.onServer(MESSENGER_EVENTS.TO_CLIENT.MESSAGE, MessengerSystem.emit);
+    alt.onServer(MESSENGER_EVENTS.TO_CLIENT.COMMANDS, MessengerSystem.setCommands);
 });