+++ src/core/plugins/core-chat/webview/utility/padNumber.ts
@@ -1,0 +1,7 @@
+export function padNumber(value: number): string {
+    if (value <= 9) {
+        return '0' + value;
+    }
+
+    return String(value);
+}