+++ src/core/shared/enums/messenger.ts
@@ -1,8 +1,9 @@
 export const MESSENGER_EVENTS = {
     TO_CLIENT: {
-        MESSAGE: 'messenger-to-client',
+        MESSAGE: 'messenger-to-client-message',
+        COMMANDS: 'messenger-to-client-commands',
     },
     TO_SERVER: {
-        MESSAGE: 'messenger-to-server',
+        MESSAGE: 'messenger-to-server-message',
     },
 };