+++ src/core/shared/interfaces/messageCommand.ts
@@ -1,0 +1,34 @@
+export type CommandCallback<T> = (player: T, ...args: any[]) => void;
+
+export interface MessageCommand<T> {
+    /**
+     * The plain text iteration of this command.
+     *
+     * @type {string}
+     * @memberof MessageCommand
+     */
+    name: string;
+
+    /**
+     * A description of this command.
+     *
+     * @type {string}
+     * @memberof MessageCommand
+     */
+    description: string;
+
+    /**
+     * An array of individual permission strings required to run this command.
+     *
+     * @type {Array<string>}
+     * @memberof MessageCommand
+     */
+    permissions: Array<string>;
+
+    /**
+     * The function to call when this command is executed by a player, or internal function.
+     *
+     * @memberof MessageCommand
+     */
+    callback: CommandCallback<T>;
+}